+++ spec/shared_examples/cnab.rb
@@ -3,7 +3,7 @@ shared_examples "cnab" do |attrs|
     {
       'id' => 1,
       'type' => 'charge',
-      'file_name' => 'FILE.TXT',
+      'filename' => 'FILE.TXT',
       'config_id' => 1,
       'charge_config_id' => 1,
     }
@@ -11,7 +11,7 @@ shared_examples "cnab" do |attrs|
 
   subject { described_class.new(attributes) }
 
-  it_behaves_like 'entity_attributes', [:id, :type, :file_name, :config_id, :charge_config_id, :status]
+  it_behaves_like 'entity_attributes', [:id, :type, :filename, :config_id, :charge_config_id, :status]
 
   describe '#charge_config_id' do
     it 'print a deprecation warning' do